+++ src/resampling/resampling_schemes/BinnedResamplings.jl
@@ -1,11 +1,17 @@
-
 """
-    AbstractBinnedResampling <: AbstractUncertainDataResampling
+    AbstractBinnedResampling
+
+Resampling schemes where data are binned.
+"""
+abstract type AbstractBinnedResampling <: AbstractUncertainDataResampling end
+    
+"""
+    AbstractBinnedUncertainValueResampling <: AbstractUncertainDataResampling
 
 Resampling schemes where data are binned and the data in each bin 
 are represented as an uncertain value.
 """
-abstract type AbstractBinnedResampling <: AbstractUncertainDataResampling end
+abstract type AbstractBinnedUncertainValueResampling <: AbstractBinnedResampling end
 
 """
     AbstractBinnedSummarisedResampling <: AbstractUncertainDataResampling
@@ -13,8 +19,9 @@ abstract type AbstractBinnedResampling <: AbstractUncertainDataResampling end
 Resampling schemes where data are binned and the data in each bin 
 are summarised to a single value (e.g. the mean).
 """
-abstract type AbstractBinnedSummarisedResampling <: AbstractUncertainDataResampling end
+abstract type AbstractBinnedSummarisedResampling <: AbstractBinnedResampling end
 
 export 
 AbstractBinnedResampling,
-AbstractBinnedSummarisedResampling
+AbstractBinnedSummarisedResampling,
+AbstractBinnedUncertainValueResampling